requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.7

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.16.0
playwright==1.40.0
aiohttp==3.9.1
fake-useragent==1.4.0

# Parsing & Data Processing
pandas==2.1.4
numpy==1.26.2
phonenumbers==8.13.26
python-dateutil==2.8.2
openpyxl==3.1.2

# Scheduling
APScheduler==3.10.4

# Instagram (optional - use with caution)
instaloader==4.10.3

# Utilities
tqdm==4.66.1
loguru==0.7.2
pydantic==2.5.3
tenacity==8.2.3

# Proxy & Rate Limiting
python-socks==2.4.3
aiohttp-socks==0.8.4
ratelimit==2.2.1

# Testing (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
